requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
passlib
bcrypt
python-multipart
aiofiles
python-jose